requirements.txt
adbutils==0.10.1
allure-pytest==2.8.16
allure-python-commons==2.8.16
amqp==1.4.9
anyjson==0.3.3
apkutils2==1.0.0
appdirs==1.4.4
Appium-Python-Client==0.46
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.2.0
baidu-aip==2.2.18.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.3.0.23
black==19.10b0
bs4==0.0.1
cached-property==1.5.2
celery==3.1.26.post2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cigam==0.0.3
Click==7.0
colorama==0.4.1
colorlog==4.0.2
concurrent-log-handler==0.9.9
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.6
deprecation==2.0.7
diff-match-patch==20181111
Django==2.0.3
django-celery==3.3.1
django-ranged-response==0.2.0
dwebsocket==0.5.12
elasticsearch==7.13.0
et-xmlfile==1.0.1
facebook-wda==1.4.0
Flask==1.1.1
flask-djcelery==0.1
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
har2case==0.3.1
HttpRunner==1.5.8
humanize==0.5.1
idna==2.6
importlib-metadata==0.23
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.10.1
jmespath==0.9.5
jsonpath==0.82
jupyter-client==6.1.7
jupyter-core==4.7.0
kafka-python==1.4.6
kombu==3.0.37
lazy-object-proxy==1.4.1
logger==1.4
loguru==0.4.1
logzero==1.5.0
lxml==4.4.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack-python==0.5.6
nb-log==4.9
odfpy==1.4.0
openpyxl==3.0.2
packaging==20.9
paramiko==2.7.2
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.0
Pillow==6.2.1
pluggy==0.13.1
progress==1.5
progressbar2==3.47.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.10.0
pycparser==2.19
pydantic==1.8.2
pyelftools==0.27
Pygments==2.4.2
pylint==2.3.1
pymongo==3.5.1
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.5
pytest==6.2.4
pytest-base-url==1.4.2
pytest-html==1.19.0
pytest-metadata==1.8.0
pytest-rerunfailures==10.1
python-dateutil==2.8.1
python-json-logger==0.1.10
python-utils==2.3.0
pytz==2019.2
PyYAML==3.12
pyzmq==18.1.0
redis==3.5.3
regex==2021.4.4
requests==2.22.0
requests-toolbelt==0.9.1
retry==0.9.2
selenium==3.141.0
sentry-sdk==0.14.4
six==1.12.0
soupsieve==2.2.1
tablib==0.14.0
texttable==0.9.1
toml==0.10.2
tomorrow3==1.1.0
tornado==5.1.1
traitlets==5.0.5
typed-ast==1.4.0
typing-extensions==3.10.0.0
uiautomator2==2.15.1.dev1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.56.0
weditor==0.6.3
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==0.6.0
